+++ docs/api.rst
@@ -262,7 +262,7 @@ skfem.assembly
 Defining a global basis
 ~~~~~~~~~~~~~~~~~~~~~~~
 
-.. autoclass:: skfem.assembly.GlobalBasis
+.. autoclass:: skfem.assembly.Basis
 
 .. autoclass:: skfem.assembly.InteriorBasis
