+++ skfem/__init__.py
@@ -13,7 +13,7 @@ __all__ = ['Mesh',
            'MeshTet',
            'MeshTri',
            'FacetBasis',
-           'GlobalBasis',
+           'Basis',
            'InteriorBasis',
            'MortarBasis',
            'asm',
