+++ skfem/assembly/basis/__init__.py
@@ -1,4 +1,4 @@
-from .global_basis import GlobalBasis
+from .basis import Basis
 from .interior_basis import InteriorBasis
 from .facet_basis import FacetBasis
 from .mortar_basis import MortarBasis
